use TestMain in trie tests
